requirements.txt
python-telegram-bot==20.0
websockets==11.0
Flask==2.3.3